HSD #18021819983: configs: socfpga: Change load address to 0x02000000
The load address is changed from 0x01000000 to 0x02000000
because the reserved space in Uboot is extended from 16MB to 32MB.
The default load address of 0x01000000 is no longer valid because
it is within the reserved region. Previously,
the reserved space is extended in Uboot for the consistent
alignment with Linux. These changes are applicable to n5x and Agilex.

Signed-off-by: Dinesh Maniyam <dinesh.maniyam@intel.com>